+++ ui/src/app/app.ts
@@ -1,11 +1,15 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
+import { Settings } from './settings/settings';
+import { SettingsService } from './settings.service';
 import { Sidebar } from './sidebar/sidebar';
 
 @Component({
-  imports: [RouterOutlet, Sidebar],
+  imports: [RouterOutlet, Sidebar, Settings],
   selector: 'app-root',
   styleUrl: './app.css',
   templateUrl: './app.html',
 })
-export class App {}
+export class App {
+  constructor(protected readonly settingsService: SettingsService) {}
+}